refactor: moved the generator out of the CLI
